创建了vo dto 对video 和 user表结构进行部分修改 写了用户的注册和登陆代码还未进行测试

+++ basic/src/main/java/org/example/basic/utils/JwtUtils.java
@@ -17,7 +17,7 @@ public class JwtUtils {
     private static final long EXPIRE_TIME = 30 * 60 * 1000;  // 30min
     private static final String JWT_TOKEN_SECRET_KEY = "douyin-simple";
 
-    public static String createJwtTokenByUser(int userId) {
+    public static String createJwtTokenByUser(long userId) {
         String secret = JWT_TOKEN_SECRET_KEY;
 
         Date date = new Date(System.currentTimeMillis() + EXPIRE_TIME);
